+++ aeolus/calc/calculus.py
@@ -174,7 +174,7 @@ def integrate(cube, coord):
     c = cube.coord(coord)
     others = [dc.name() for dc in cube.dim_coords if cube.coord_dims(dc) != cube.coord_dims(c)]
     dim = cube.coord_dims(c)[0]
-    data = np.trapz(cube.data, c.points, axis=dim)
+    data = np.trapz(cube.core_data(), c.points, axis=dim)
     res = next(cube.slices(others)).copy(data=data)
     res.units = cube.units * c.units
     res.remove_coord(c)